views: Move json_refer_friend to invite.py.
