+++ zproject/legacy_urls.py
@@ -8,6 +8,7 @@ legacy_urls = [
     url(r'^json/make_stream_private$',      'zerver.views.streams.json_make_stream_private'),
     url(r'^json/invite_users$',             'zerver.views.invite.json_invite_users'),
     url(r'^json/bulk_invite_users$',        'zerver.views.invite.json_bulk_invite_users'),
+    url(r'^json/refer_friend$',             'zerver.views.invite.json_refer_friend'),
     url(r'^json/settings/change$',          'zerver.views.user_settings.json_change_settings'),
     url(r'^json/notify_settings/change$',   'zerver.views.user_settings.json_change_notify_settings'),
     url(r'^json/ui_settings/change$',       'zerver.views.user_settings.json_change_ui_settings'),
@@ -26,7 +27,6 @@ legacy_urls = [
     url(r'^json/messages_in_narrow$',       'zerver.views.messages.json_messages_in_narrow'),
     url(r'^json/update_message$',           'zerver.views.messages.json_update_message'),
     url(r'^json/fetch_raw_message$',        'zerver.views.messages.json_fetch_raw_message'),
-    url(r'^json/refer_friend$',             'zerver.views.json_refer_friend'),
     url(r'^json/set_muted_topics$',         'zerver.views.json_set_muted_topics'),
     url(r'^json/set_avatar$',               'zerver.views.user_settings.json_set_avatar'),
     url(r'^json/time_setting$',             'zerver.views.user_settings.json_time_setting'),
